move threadsafe datastructures to internal and fix EngineStats for JS
