+++ core/src/commonMain/kotlin/com/littlekt/graphics/Texture.kt
@@ -235,6 +235,7 @@ interface Texture : Releasable {
         val buffer = commandEncoder.finish()
         device.queue.submit(buffer)
         commandEncoder.release()
+        mesh.release()
         views.forEach { it.release() }
         sampler.release()
         shader.release()
